[ADD] Constants/Enum/ 폴더링, EventType추가
