+++ PPAK_CVS/Configuration/Constants/Enum/CvsType.swift
@@ -1,5 +1,5 @@
 //
-//  Enum.swift
+//  CvsType.swift
 //  PPAK_CVS
 //
 //  Created by hyeonseok on 2022/10/03.
@@ -66,22 +66,3 @@ enum CVSType {
   }
 }
 
-/// Description
-/// 이벤트의 종류를 가리키는 Enum Type
-enum EventType {
-  
-  case onePlusOne
-  case twoPlusOne
-  case allEnText
-  case allKrText
-  
-  var eventTitle: String {
-    switch self {
-    case .onePlusOne: return "1 + 1"
-    case .twoPlusOne: return "2 + 1"
-    case .allEnText: return "All"
-    case .allKrText: return "전체"
-    }
-  }
-  
-}
